[Feat] 회전 무기 구현

+++ Rogue-Lite/Assets/Scripts/Character/Status.cs
@@ -12,7 +12,7 @@ public class Status : MonoBehaviour
         return _status.Speed;
     }
 
-    public float GetDamage()
+    public int GetDamage()
     {
         return _status.AttackDamage;
     }
